// frontend/src/api/model/shoppingProductRequest.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ShoppingProductRequest {
  /**
   * Name of shopping item
   * @minLength 1
   */
  name: string;
  /** Description of shopping item */
  description?: string;
  /** Number of items in the shopping list */
  quantity?: number;
  /** Has the item been purchased or not. Optional defaults to false */
  purchased?: boolean;
}