// frontend/src/api/model/shoppingProductResponse.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ShoppingProductResponse {
  id?: number;
  name?: string;
  description?: string;
  quantity?: number;
  purchased?: boolean;
  createdAt?: string;
  updatedAt?: string;
}